Extract TokenValueStat component
