+++ packages/ui/src/common/components/statistics/index.tsx
@@ -1,2 +1,3 @@
 export * from './Statistics'
 export * from './StatisticItem'
+export * from './TokenValueStat'
